feat(auth): support SAML strategy
